Fully implement MVP as gui pattern
In this commit a lot happened. I deleted the requirement for
the boost library since it isnt used yet. Furthermore i created
a pretty simple window (but it is useful thouh). It already
has an open dialogue. And no everything fits the MVP pattern. So there is
a Model, View and a Presenter. File can already get loaded.